Moved the 'device' and 'grid' subpackages out of the 'core' subpackage to the quatrex top level.
Also adjusted imports accordingly.

Feels a bit more natural this way. There is an argument to be had about whether 'core' should even exist as it is not very descriptive :skull:
