+++ src/quatrex/device/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2024 ETH Zurich and the authors of the quatrex package.
 
-from quatrex.core.device.contact import Contact
-from quatrex.core.device.device import Device
+from quatrex.device.contact import Contact
+from quatrex.device.device import Device
 
 __all__ = ["Contact", "Device"]
