+++ src/quatrex/device/contact.py
