+++ src/quatrex/grid/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2024 ETH Zurich and the authors of the quatrex package.
 
-from quatrex.core.grid.energies import get_electron_energies
-from quatrex.core.grid.kpoints import monkhorst_pack
+from quatrex.grid.energies import get_electron_energies
+from quatrex.grid.kpoints import monkhorst_pack
 
 __all__ = ["get_electron_energies", "monkhorst_pack"]
